Add gabby.pkl and matthew.pkl so that MongoDB tags are correct
Currently, MongoDB tags are based on gabby and matthew, not dataset1 and dataset2. In order to reduce confusion, add copies with the original names so that scraping can continue as normal. Eventually, change this so that the tag 'gabby' is completely replaced with 'dataset1' and the tag 'matthew' is completely replaced with 'dataset2'